chore: move simple scenario into own file
I refactor a bit and renamed some things to make it more clear where things come from since a lot of names are shared
